Fixed issue for localized domains using the same path alias
If you have two localizations, say a US site and a European site with
the same path that needs to be redirected say (/About-Us-> /AboutUs) it
would fail because the query was keying off the URL path only returning
2 results.  With this change the match query keys on the URL path and
the hostname ensuring 1 result.

I've also modified the following if condition since we no longer have to
test if the hostname matches the inbound hostname since we did that in
the match query.